fix: align multiclass slot test fixture
